Merge: MarketPriceTable-task3 머지
