+++ src/components/MainBoard/Chip.css
