+++ src/components/MainBoard/ChipBox.css
