+++ src/components/MarketPriceTable/constants/currencyUnits.js
@@ -1,0 +1,6 @@
+const currencyUnits = {
+  krw: '￦',
+  usd: '$',
+};
+
+export default currencyUnits;
